Remove magic number
