+++ constants_local.xml
@@ -21,8 +21,8 @@
         <NUM_FINAL_BLOCK_PER_POW>50</NUM_FINAL_BLOCK_PER_POW>
         <NUM_DS_KEEP_TX_BODY>5</NUM_DS_KEEP_TX_BODY>
         <MAXMESSAGE>32</MAXMESSAGE>
-        <MAXSUBMITTXNPERNODE>10</MAXSUBMITTXNPERNODE>
-        <MICROBLOCK_GAS_LIMIT>500</MICROBLOCK_GAS_LIMIT>
+        <MAXSUBMITTXNPERNODE>10000</MAXSUBMITTXNPERNODE>
+        <MICROBLOCK_GAS_LIMIT>500000</MICROBLOCK_GAS_LIMIT>
         <NEW_NODE_POW_DELAY>5</NEW_NODE_POW_DELAY>
         <POST_VIEWCHANGE_BUFFER>5</POST_VIEWCHANGE_BUFFER>
         <CONTRACT_CREATE_GAS>50</CONTRACT_CREATE_GAS>
@@ -33,7 +33,9 @@
         <DEBUG_LEVEL>3</DEBUG_LEVEL>
         <BROADCAST_INTERVAL>60</BROADCAST_INTERVAL>
         <BROADCAST_EXPIRY>600</BROADCAST_EXPIRY>
-        <TX_DISTRIBUTE_TIME_IN_MS>10000</TX_DISTRIBUTE_TIME_IN_MS>
+        <TX_DISTRIBUTE_TIME_IN_MS>3000</TX_DISTRIBUTE_TIME_IN_MS>
+		<NUM_NODES_TO_SEND_LOOKUP>3</NUM_NODES_TO_SEND_LOOKUP>
+		<NUM_TXN_TO_SEND_PER_ACCOUNT>4000</NUM_TXN_TO_SEND_PER_ACCOUNT>
     </constants>
     <options>
         <TEST_NET_MODE>false</TEST_NET_MODE>
